requirements.txt
torch
torchvision
pandas
numpy
opencv-python
matplotlib
pillow
tqdm
